Resolve "Enable all clang-tidy bugprone checks"

Closes #831

See merge request project-march/march!877

+++ ros1/src/data/march_realsense_reader/src/pointcloud_processor/preprocessor.cpp
@@ -136,7 +136,7 @@ void NormalsPreprocessor::getDownsamplingParameters()
         random_filter = yaml_utilities::grabParameter<bool>(
             downsampling_parameters, "random_filter");
         if (voxel_grid_filter) {
-            leaf_size = yaml_utilities::grabParameter<double>(
+            leaf_size = yaml_utilities::grabParameter<float>(
                 downsampling_parameters, "leaf_size");
         } else if (random_filter) {
             remaining_points = yaml_utilities::grabParameter<int>(
